move snapshotting logic to Checkpoints
